Increase the sliding menu
