+++ src/components/GlobalHeader/index.less
@@ -3,7 +3,7 @@
 .header {
   padding: 0 12px 0 0;
   background: #fff;
-  box-shadow: 0 1px 4px rgba(0, 21, 41, .08);
+  box-shadow: 0 1px 4px rgba(0, 21, 41, 0.08);
   position: relative;
 }
 
@@ -26,16 +26,10 @@ i.trigger {
   font-size: 20px;
   line-height: 64px;
   cursor: pointer;
-  transition: all .3s;
+  transition: all 0.3s;
   padding: 0 24px;
   &:hover {
     background: @primary-1;
-  }
-}
-
-@media screen and (max-width: @screen-xs) {
-  .trigger {
-    display: none;
   }
 }
 
@@ -46,7 +40,7 @@ i.trigger {
     cursor: pointer;
     padding: 0 12px;
     display: inline-block;
-    transition: all .3s;
+    transition: all 0.3s;
     height: 100%;
     > i {
       font-size: 16px;
@@ -68,7 +62,7 @@ i.trigger {
     .avatar {
       margin: 20px 8px 20px 0;
       color: @primary-color;
-      background: rgba(255, 255, 255, .85);
+      background: rgba(255, 255, 255, 0.85);
       vertical-align: middle;
     }
   }
