+++ src/components/SiderMenu/index.less
@@ -5,7 +5,7 @@
   position: relative;
   line-height: 64px;
   padding-left: (@menu-collapsed-width - 32px) / 2;
-  transition: all .3s;
+  transition: all 0.3s;
   background: #002140;
   overflow: hidden;
   img {
@@ -19,14 +19,28 @@
     vertical-align: middle;
     font-size: 20px;
     margin: 0 0 0 12px;
-    font-family: 'Myriad Pro', 'Helvetica Neue', Arial, Helvetica, sans-serif;
+    font-family: "Myriad Pro", "Helvetica Neue", Arial, Helvetica, sans-serif;
     font-weight: 600;
   }
 }
 
 .sider {
   min-height: 100vh;
-  box-shadow: 2px 0 6px rgba(0, 21, 41, .35);
+  box-shadow: 2px 0 6px rgba(0, 21, 41, 0.35);
   position: relative;
   z-index: 10;
+}
+
+.navSidebar {
+  position: fixed;
+  left: -258px;
+  transition: left .3s;
+  top: 0;
+  z-index: 10;
+  height: 100%;
+  .over {
+    height: 100%;
+    width: 100%;
+    position: fixed;
+  }
 }
